applied authorization to get, update, patch, delete handlers for entry

+++ telegram_bot/routers/entry_bot/entry_delete.py
@@ -6,12 +6,34 @@ from .entry_router import router
 import httpx
 
 API_URL = "http://127.0.0.1:8000/entries/{entry_id}"
+API_GET_URL = "http://127.0.0.1:8000/token/{telegram_id}"
+API_TOKEN_URL = "http://127.0.0.1:8000/auth/validate"
 
 class DeleteEntryState(StatesGroup):
     waiting_id = State()
 
 @router.message(Command("delete_entry"))
 async def get_id(message: types.Message, state: FSMContext):
+    telegram_id = message.from_user.id
+    async with httpx.AsyncClient() as client:
+        response = await client.get(API_GET_URL.format(telegram_id=telegram_id))
+
+    if response.status_code != 200:
+        await message.answer(f"User with telegram id {telegram_id} unauthorized. Use command /login for authorize")
+        await state.clear()
+        return
+
+    token = response.json().get("access_token")
+
+    async with httpx.AsyncClient() as client:
+        response = await client.get(API_TOKEN_URL, headers={"Authorization": f"Bearer {token}"})
+
+    if response.status_code != 200:
+        await message.answer(f"User didn't authorize. Use command /login for authorize")
+        await state.clear()
+        return
+
+    await state.update_data(token=token)
     await message.answer("Enter the id of entry:")
     await state.set_state(DeleteEntryState.waiting_id)
 
@@ -25,8 +47,10 @@ async def delete_entry(message: types.Message, state: FSMContext):
         await message.answer("Enter a integer number")
         return
 
+    data = await state.get_data()
+    token = data.pop("token")
     async with httpx.AsyncClient() as client:
-        response = await client.delete(API_URL.format(entry_id=entry_id))
+        response = await client.delete(API_URL.format(entry_id=entry_id), headers={"Authorization": f"Bearer {token}"})
 
     if response.status_code == 200:
         await message.answer("Entry deleted successfully ✅")
